grumbles: split host filter code
